+++ NEWS.md
@@ -7,3 +7,6 @@
 # onehot 0.1.1.9000
 
 * Fixed a bug in onehot.summary switch statement. Using proper default now.
+* Added support for predicting a sparse matrix.
+* Added multicore support for predicting dense matrices using openMP.
+* New argument, addNA, adds indicator columns for all input columns.